+++ CHANGELOG.md
@@ -4,7 +4,18 @@ This changelog follows the specifications detailed in: [Keep a Changelog](https:
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
 
-## Version 0.6.18 - Unreleased
+## Version 0.6.19 - Unreleased
+
+### Added
+
+* Support for numpy 2.0 on Python 3.9+
+
+
+## Version 0.6.18 - Released 2024-03-19
+
+### Added
+
+* Add memmap support to ``Stitcher`` in numpy mode.
 
 
 ## Version 0.6.17 - Released 2024-02-01